Renaming the noos packages to embedded
