+++ driver/embedded/event.go
@@ -1,4 +1,4 @@
-package noos
+package embedded
 
 type Event interface {
 	isEvent()
